requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
celery==5.3.6
redis==5.0.1
gspread==6.1.2
gspread-formatting==1.2.0
slack-sdk==3.26.2
PyGithub==2.1.1
python-dotenv==1.0.0

